FInanceilmApiService/requirements.txt
typing-extensions>=4.9.0
chromadb==1.0.16
langchain==0.2.16
langchain-openai==0.1.17
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1
openai>=1.3.0
tiktoken>=0.7.0
numpy>=1.26.0
chromadb>=1.0.15
requests>=2.31.0
gunicorn>=21.2.0